vision_action_agent/requirements.txt
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.1
pyautogui==0.9.54
pytesseract==0.3.10
easyocr==1.7.0
pynput==1.7.6
psutil==5.9.5
pyyaml==6.0.1
configparser==6.0.0
loguru==0.7.2
typing-extensions==4.8.0
dataclasses-json==0.6.1
click==8.1.7